chore (refs T22614): move rest core Components to client/js.
