+++ client/js/components/core/CopyPasteButton.vue
