+++ client/js/components/core/DpSkeletonBox.vue
