+++ client/js/components/core/Obscure.vue
