+++ demosplan/DemosPlanNewsBundle/Resources/client/js/bundles/newsAdminNew.js
@@ -11,7 +11,7 @@
  * This is the entry point for news_admin_new.html.twig
  */
 
-import DpChangeStateAtDate from '@DemosPlanCoreBundle/components/DpChangeStateAtDate'
+import DpChangeStateAtDate from '@DpJs/components/core/DpChangeStateAtDate'
 import DpEditor from '@DpJs/components/core/DpEditor/DpEditor'
 import DpUploadFiles from '@DpJs/components/core/DpUpload/DpUploadFiles'
 import dpValidate from '@DpJs/lib/validation/dpValidate'
