+++ demosplan/DemosPlanProcedureBundle/Resources/client/js/components/AdministrationImport/ExcelImport/ExcelImportErrors.vue
@@ -75,7 +75,7 @@
 <script>
 import DpCheckbox from '@DpJs/components/core/form/DpCheckbox'
 import DpInlineNotification from '@DpJs/components/core/DpInlineNotification'
-import DpProgressBar from '@DemosPlanCoreBundle/components/DpProgressBar'
+import DpProgressBar from '@DpJs/components/core/DpProgressBar'
 
 export default {
   name: 'ExcelImportErrors',
