+++ demosplan/DemosPlanUserBundle/Resources/client/js/components/DpOrganisationList.vue
@@ -157,7 +157,7 @@ import DpCheckboxGroup from '@DpJs/components/core/DpCheckboxGroup'
 import DpOrganisationListItem from './DpOrganisationListItem'
 import DpSearchField from '@DpJs/components/core/form/DpSearchField'
 import dpSelectAllMixin from '@DpJs/lib/utils/dpSelectAllMixin'
-import DpSkeletonBox from '@DemosPlanCoreBundle/components/DpSkeletonBox'
+import DpSkeletonBox from '@DpJs/components/core/DpSkeletonBox'
 import DpSlidingPagination from '@DpJs/components/core/DpSlidingPagination'
 import hasOwnProp from '@DpJs/lib/utils/hasOwnProp'
